GX: migrate the namespace out of C mode (18 files) — phase 1 of #821
First slice of phase 1 in notes/plan-cpp-language-mode.md. GX is a namespace,
not a class: no `this`, no vtable, no header, no includers. So these are the
lowest-risk migrations in the tree -- nothing outside each file can shift -- and
they are the right place to prove the pipeline before touching a class.

`void _ZN2GX6DispOnEv(void)` becomes `namespace GX { void DispOn() }`. The
compiler now mangles the symbol instead of the author spelling it, which is the
actual complaint in #821. Bodies are untouched; function `extern`s gained C
linkage so their unmangled names still resolve.

Deliberately NOT done here, to keep one codegen variable per verification:

  * callee declarations are left as-is, raw mangled names included. Repairing
    those is per-reference work (phase 3). GX::SetBankForTex is declared taking
    u32 while its mangled name says `Et` (unsigned short) -- a real defect, and
    a separate one.
  * the local `struct VramReg` in SetBankForSubBG stays. Shadow declarations are
    phase 4; the counts do not move either way here.

That completes GX: all 22 of its files are now real C++, 18 of them by this
commit.

VERIFICATION
  * every one byte-verifies under the pinned compiler:
    build_pin.verify -> (True, '2004/b56') for all 18, individually.
  * port_refcheck caught a stale port/ manifest reference to
    _ZN2GX11LoadTexPlttEPKvjj.c -- exactly the .c-to-.cpp trap MERGE.md warns
    about and that `validate` cannot see. Fixed; 394 references now resolve.
  * prepush_attribution: 0 changed, 0 lost. The #938 same-stem rescue covers
    .c -> .cpp, and no file moves directory here.
  * layout_check: clean.
  * delinks paths follow the extension. Hand-applied rather than regenerated,
    because `enroll.py` also wants two changes unrelated to this work (it drops
    the NONMATCHING _ZN9ActorBaseC1Ev.cpp entry and adds a rombytes entry for
    func_ov007_020cbbb0.c). Diffed against a full regeneration to confirm these
    18 lines are exactly what it would produce.
  * rombuild.py's 106/106 gate could not run locally -- no ROM dump on this
    machine -- so the module-fidelity verdict is CI's. No header is touched, so
    the eligible.py blast-radius bracket does not apply.
  * langmode ratchet: unmigrated 1359 -> 1341, .c-extension 1044 -> 1026.
    Baseline re-banked in this commit, as the workflow header instructs.
