+++ config/arm9/delinks.txt
@@ -4312,7 +4312,7 @@ src/_ZN5Stage12SetVramBanksEv.c:
     complete
     .text start:0x0202e73c end:0x0202e78c
 
-src/_ZN2GX15DisableAllBanksEv.c:
+src/_ZN2GX15DisableAllBanksEv.cpp:
     complete
     .text start:0x0202e78c end:0x0202e7d4
 
@@ -8668,11 +8668,11 @@ src/_ZN3GXS15SetGraphicsModeEi.cpp:
     complete
     .text start:0x02053abc end:0x02053ad8
 
-src/_ZN2GX15SetGraphicsModeEiii.c:
+src/_ZN2GX15SetGraphicsModeEiii.cpp:
     complete
     .text start:0x02053ad8 end:0x02053b4c
 
-src/_ZN2GX6DispOnEv.c:
+src/_ZN2GX6DispOnEv.cpp:
     complete
     .text start:0x02053b4c end:0x02053b98
 
@@ -8816,19 +8816,19 @@ src/func_020541b8.c:
     complete
     .text start:0x020541b8 end:0x020541f0
 
-src/_ZN2GX23SetBankForSubOBJExtPlttEt.c:
+src/_ZN2GX23SetBankForSubOBJExtPlttEt.cpp:
     complete
     .text start:0x020541f0 end:0x0205427c
 
-src/_ZN2GX22SetBankForSubBGExtPlttEt.c:
+src/_ZN2GX22SetBankForSubBGExtPlttEt.cpp:
     complete
     .text start:0x0205427c end:0x02054308
 
-src/_ZN2GX16SetBankForSubOBJEt.c:
+src/_ZN2GX16SetBankForSubOBJEt.cpp:
     complete
     .text start:0x02054308 end:0x02054384
 
-src/_ZN2GX15SetBankForSubBGEt.c:
+src/_ZN2GX15SetBankForSubBGEt.cpp:
     complete
     .text start:0x02054384 end:0x02054430
 
@@ -8836,7 +8836,7 @@ src/func_02054430.c:
     complete
     .text start:0x02054430 end:0x02054450
 
-src/_ZN2GX17SetBankForTexPlttEt.c:
+src/_ZN2GX17SetBankForTexPlttEt.cpp:
     complete
     .text start:0x02054450 end:0x02054544
 
@@ -9063,7 +9063,7 @@ src/_ZN3GXS11LoadBG1CharEPKvjj.c:
     complete
     .text start:0x02056134 end:0x02056194
 
-src/_ZN2GX11LoadBG1CharEPKvjj.c:
+src/_ZN2GX11LoadBG1CharEPKvjj.cpp:
     complete
     .text start:0x02056194 end:0x020561f4
 
@@ -9071,7 +9071,7 @@ src/_ZN3GXS11LoadBG0CharEPKvjj.c:
     complete
     .text start:0x020561f4 end:0x02056254
 
-src/_ZN2GX11LoadBG0CharEPKvjj.c:
+src/_ZN2GX11LoadBG0CharEPKvjj.cpp:
     complete
     .text start:0x02056254 end:0x020562b4
 
@@ -9111,7 +9111,7 @@ src/func_020565b4.c:
     complete
     .text start:0x020565b4 end:0x02056614
 
-src/_ZN2GX7LoadOBJEPKvjj.c:
+src/_ZN2GX7LoadOBJEPKvjj.cpp:
     complete
     .text start:0x02056614 end:0x02056674
 
@@ -9127,7 +9127,7 @@ src/_ZN3GXS11LoadOBJPlttEPKvjj.c:
     complete
     .text start:0x02056738 end:0x020567a0
 
-src/_ZN2GX11LoadOBJPlttEPKvjj.c:
+src/_ZN2GX11LoadOBJPlttEPKvjj.cpp:
     complete
     .text start:0x020567a0 end:0x02056808
 
@@ -9135,23 +9135,23 @@ src/_ZN3GXS10LoadBGPlttEPKvjj.c:
     complete
     .text start:0x02056808 end:0x02056870
 
-src/_ZN2GX10LoadBGPlttEPKvjj.c:
+src/_ZN2GX10LoadBGPlttEPKvjj.cpp:
     complete
     .text start:0x02056870 end:0x020568cc
 
-src/_ZN2GX14EndLoadTexPlttEv.c:
+src/_ZN2GX14EndLoadTexPlttEv.cpp:
     complete
     .text start:0x020568cc end:0x02056924
 
-src/_ZN2GX11LoadTexPlttEPKvjj.c:
+src/_ZN2GX11LoadTexPlttEPKvjj.cpp:
     complete
     .text start:0x02056924 end:0x02056998
 
-src/_ZN2GX16BeginLoadTexPlttEv.c:
+src/_ZN2GX16BeginLoadTexPlttEv.cpp:
     complete
     .text start:0x02056998 end:0x020569e0
 
-src/_ZN2GX10EndLoadTexEv.c:
+src/_ZN2GX10EndLoadTexEv.cpp:
     complete
     .text start:0x020569e0 end:0x02056a50
 
@@ -9159,7 +9159,7 @@ src/_ZN2GX7LoadTexEPKvjj.cpp:
     complete
     .text start:0x02056a50 end:0x02056b9c
 
-src/_ZN2GX12BeginLoadTexEv.c:
+src/_ZN2GX12BeginLoadTexEv.cpp:
     complete
     .text start:0x02056b9c end:0x02056c1c
