--- src/_ZN2GX16BeginLoadTexPlttEv.c
@@ -1,12 +1,0 @@
-#include "types.h"
-extern int func_02054168(void);
-extern int data_020a60b4;
-extern u16 data_02086314[];
-extern int data_020a60b0;
-
-void _ZN2GX16BeginLoadTexPlttEv(void)
-{
-    int r = func_02054168();
-    data_020a60b4 = r;
-    data_020a60b0 = (int)data_02086314[r >> 4] << 12;
-}
