+++ src/_ZN2GX17SetBankForTexPlttEt.cpp
@@ -1,8 +1,21 @@
+//cpp
+// @symbol _ZN2GX17SetBankForTexPlttEt
+/* GX::SetBankForTexPltt() -- language-mode migration only (phase 1 of
+ * notes/plan-cpp-language-mode.md). GX is a namespace, not a class: no
+ * `this`, no vtable, no layout, so nothing outside this file can shift.
+ * The body and every callee declaration are unchanged from the C version;
+ * the symbol is now mangled by the compiler instead of spelled by hand.
+ */
 #include "types.h"
 struct VramReg { u16 w0; u16 pad[4]; u16 fa; };
 extern struct VramReg data_020a6088;
+
+extern "C" {
 extern void Vram__Map(u16 bits);
-void _ZN2GX17SetBankForTexPlttEt(u16 x){
+}
+
+namespace GX {
+void SetBankForTexPltt(u16 x){
     data_020a6088.w0 = ~x & (data_020a6088.w0 | data_020a6088.fa);
     data_020a6088.fa = x;
     switch (x) {
@@ -24,4 +37,5 @@ void _ZN2GX17SetBankForTexPlttEt(u16 x){
         break;
     }
     Vram__Map(data_020a6088.w0);
+}
 }
